chore(ref): Improve API and refactor users management code
