+++ authx/routers/auth.py
@@ -3,7 +3,7 @@ from typing import Callable, Dict
 from fastapi import APIRouter, Body, Depends, Request, Response
 from fastapi.exceptions import HTTPException
 
-from authx.api import UsersRepo
+from authx.backend import UsersRepo
 from authx.core.jwt import JWTBackend
 from authx.core.user import User
 from authx.services import AuthService
