+++ authx/routers/social.py
@@ -5,7 +5,7 @@ from typing import Iterable, Optional
 from fastapi import APIRouter, HTTPException, Request
 from fastapi.responses import HTMLResponse, RedirectResponse
 
-from authx.api import UsersRepo
+from authx.backend import UsersRepo
 from authx.core.jwt import JWTBackend
 from authx.errors import SocialException
 from authx.services import SocialService
