+++ authx/services/search.py
@@ -2,7 +2,7 @@ from typing import Optional
 
 from fastapi import HTTPException
 
-from authx.api import UsersRepo
+from authx.backend import UsersRepo
 from authx.models.user import UserPrivateInfo
