+++ authx/services/social.py
@@ -4,7 +4,7 @@ from typing import Dict, Optional, Tuple
 import jwt
 from httpx import AsyncClient
 
-from authx.api import UsersRepo
+from authx.backend import UsersRepo
 from authx.core.jwt import JWTBackend
 from authx.errors import SocialException
 from authx.models.social import SocialInCreate
